+++ internal/interface/cli/sbi/sbi.go
@@ -21,6 +21,7 @@ func NewSBICommand() *cobra.Command {
 	cmd.AddCommand(NewSBIShowCommand())
 	cmd.AddCommand(NewSBIResetCommand())
 	cmd.AddCommand(NewSBIHistoryCommand())
+	cmd.AddCommand(NewSBIReviewCommand())
 
 	return cmd
 }